Works/RAG_bot_ask_me_from_PDF/requirements.txt

streamlit
langchain
langchain-openai
faiss-cpu
python-dotenv
PyPDF2
tiktoken
